wpgsd: Weighted Parametric Group Sequential Design

Adjusted inference for weighted parametric group sequential design. Weighted parametric group sequential design (WPGSD) by Anderson et al. (2022) doi:10.1002/bimj.202100085 allows one to take advantage of the known correlation structure in constructing efficacy bounds to control family-wise error rate (FWER) for a group sequential design. Here, correlation may be due to common observations in nested populations, due to common observations in overlapping populations, or due to common observations in the control arm.

See also

Author

Maintainer: Yujie Zhao yujie.zhao@merck.com

Authors:

Keaven Anderson keaven_anderson@merck.com

Zifang Guo zifang.guo@merck.com

Jing Zhao jing_zhaox@merck.com

Linda Sun linda_sun@merck.com

Yi Cui yi.cui@merck.com

Larry Leon larry.leon2@merck.com

Other contributors:

Merck Sharp & Dohme Corp [copyright holder]