Generate table of event counts from ADSL and ADTTE datasets

Generate table of event counts from ADSL and ADTTE datasets

Usage

generate_event_table(paths, h_select, adsl_name, adtte_name, key_var, cnsr_var)

Arguments

paths

A vector of paths for analysis datasets. Length should be equal to the number of analyses completed.

h_select

Selection criterion for each hypothesis. Should be a tibble containing 2 columns: Hypothesis and Crit.

adsl_name

SAS dataset name for subject-level analysis data. Usually it is "adsl".

adtte_name

SAS dataset name for time-to-event analysis data. Usually it is "adtte".

key_var

Key variable to join the adsl and adtte datasets. For example, "USUBJID" or "SUBJID".

cnsr_var

Variable to indicate censoring (1 = censor; 0 = event). For example, "CNSR".

Value

A list with two components:

event: an event count table as input for generate_bounds().

dsets: analysis datasets of each hypothesis.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

paths <- system.file("extdata/", package = "wpgsd")

# Selection criteria for each hypothesis
h_select <- tibble::tribble(
  ~Hypothesis, ~Crit,
  1, "PARAMCD == 'OS' & TRT01P %in% c('Xanomeline High Dose', 'Placebo')",
  2, "PARAMCD == 'OS' & TRT01P %in% c('Xanomeline Low Dose', 'Placebo')"
)

event <- generate_event_table(paths, h_select,
  adsl_name = "adsl", adtte_name = "adtte",
  key_var = "USUBJID", cnsr_var = "CNSR"
)$event

event %>%
  kableExtra::kbl(caption = "Event Count - Compute from SAS Datasets Example", align = "l") %>%
  kableExtra::kable_classic_2(full_width = FALSE)
#> <table class=" lightable-classic-2" style='font-family: "Arial Narrow", "Source Sans Pro", sans-serif; width: auto !important; margin-left: auto; margin-right: auto;'>
#> <caption>Event Count - Compute from SAS Datasets Example</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> H1 </th>
#>    <th style="text-align:left;"> H2 </th>
#>    <th style="text-align:left;"> Analysis </th>
#>    <th style="text-align:left;"> Event </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:left;"> 66 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:left;"> 59 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:left;"> 45 </td>
#>   </tr>
#> </tbody>
#> </table>